+++ src/js/chameleon-wiki.js
@@ -1,5 +1,8 @@
 require("./wiki/anchor");
 require("./wiki/edit-section");
 require("./wiki/login-modal");
+require("./wiki/preferences");
 require("./wiki/toc");
+require("./wiki/class-name-mapping");
+
 require("./wiki/migrations");