+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- icon: add bootstrap-icons
+
+## Changed
+
+- bootstrap: update to 4.5.0
+- js: `jquery.js` and `bootstrap.js` are not bundled anymore **BREAKING CHANGE**
+- js,icon: optimize size
+- js,css: new megamenu design
+
+## Removed
+
+- icon: remove _remix icon_ **BREAKING CHANGE**
+
+## Fixed
+
+- css: table colors in dark mode
+
 ## [2.0.0] - 2020-03-04
 
 ### Added